Breaks up event detail stories into individual files and switches to args.
